improve the crud logic with middle utils functions

+++ user_crud.py
@@ -15,29 +15,36 @@ class UserRepo:
             'name': str,
             'email': str,
         }
-    ):
+    ) -> str:
         user_obj = User(name=user['name'], email=user['email'])
         self.db.add(user_obj)
         self.db.commit()
         self.db.refresh(user_obj)
         return f"Added user {user_obj.name} with email {user_obj.email}"
 
-    def fetch_by_id(self, user_id: int):
-        return self.db.query(User).filter(User.id == user_id).first()
+    def fetch_by_id(self, user_id: int) -> User | str:
+        user = self.db.query(User).filter(User.id == user_id).first()
+        if not user:
+            return f"User with id {user_id} not found"
+        return user
 
     def fetch_all(self, skip: int = 0, limit: int = 10):
         return self.db.query(User).offset(skip).limit(limit).all()
 
-    def delete(self, user_id: int):
+    def delete(self, user_id: int) -> str:
         user = self.fetch_by_id(user_id)
+        if not isinstance(user, User):
+            return f"Coudn't delete user: User with id {user_id} not found"
         self.db.delete(user)
         self.db.commit()
         return f"Deleted user {user.name} with email {user.email}"
 
-    def update(self, user_id: int, user: dict):
+    def update(self, user_id: int, user: dict) -> str:
         user_obj = self.fetch_by_id(user_id)
-        user_obj.name = user['name'] if user['name'] else user.name
-        user_obj.email = user['email'] if user['email'] else user.email
+        if not isinstance(user_obj, User):
+            return f"Coudn't update user: User with id {user_id} not found"
+        for key, value in user.items():
+            setattr(user_obj, key, value)
         self.db.commit()
         self.db.refresh(user_obj)
         return f"Updated user {user_obj.name} with email {user_obj.email}"
